+++ src/main/java/cc/kertaskerja/tppkepegawaian/opd/domain/OpdService.java
@@ -30,6 +30,10 @@ public class OpdService {
 	}
 	
 	public void hapusOpd(String kodeOpd) {
+	if (!opdRepository.existsByKodeOpd(kodeOpd)) {
+	    throw new OpdNotFoundException(kodeOpd);
+	} 
+	   
         opdRepository.deleteByKodeOpd(kodeOpd);
     }
 }
